java/src/test/java/com/yourtion/leetcode/easy/c9009/SolutionTest.java
package com.yourtion.leetcode.easy.c9009;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.params.provider.Arguments.arguments;

@DisplayName("面试题 01.04. 回文排列")
class SolutionTest {

    static Stream<Arguments> testDataProvider() {
        return Stream.of(
                arguments("tactcoa", true),
                arguments("ab", false)
        );
    }

    @ParameterizedTest()
    @MethodSource("testDataProvider")
    void canPermutePalindrome(String source, boolean res) {
        System.out.printf("runTest: %s , res: %b", source, res);
        Assertions.assertEquals(res, new Solution().canPermutePalindrome(source));
    }
}